requirements.txt
streamlit
pandas
openai
matplotlib
seaborn
python-dotenv